+++ assets/spark-patch/Celeborn_Dynamic_Allocation_spark2.patch
@@ -43,7 +43,7 @@ diff --git a/core/src/main/scala/org/apache/spark/ExecutorAllocationManager.scal
 +      !Utils.isCelebornEnabled(conf) && !testing) {
 +      throw new SparkException("Dynamic allocation of executors requires the external or remote " +
 +        "shuffle service. You may enable this through spark.shuffle.service.enabled or " +
-+        "set spark.shuffle.manager=org.apache.spark.shuffle.celeborn.CelebornShuffleManager.")
++        "set spark.shuffle.manager=org.apache.spark.shuffle.celeborn.SparkShuffleManager.")
      }
      if (tasksPerExecutorForFullParallelism == 0) {
        throw new SparkException("spark.executor.cores must not be < spark.task.cpus.")
