+++ toolkit/scripts/genConfs.py
@@ -94,7 +94,7 @@ def set_skew_join_confs(conf):
 
 
 def set_rss_confs(conf, replicate=False):
-    conf["spark.shuffle.manager"] = "org.apache.spark.shuffle.celeborn.CelebornShuffleManager"
+    conf["spark.shuffle.manager"] = "org.apache.spark.shuffle.celeborn.SparkShuffleManager"
     conf["spark.serializer"] = "org.apache.spark.serializer.KryoSerializer"
     conf["spark.celeborn.master.endpoints"] = "master-1-1:9097"
     conf["spark.shuffle.service.enabled"] = "false"
